requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
chromadb==0.3.27


streamlit==1.26.0
python-dotenv==1.0.1


python-docx==0.8.11
docx2txt==0.8


torch==2.2.2+cpu
sentence-transformers==2.2.2
transformers==4.38.0 
huggingface-hub==0.20.3


chromadb==0.3.27

pydantic==1.9.0


tqdm==4.66.4
python-magic==0.4.27
typing-extensions>=4.8.0,<5


langchain==0.0.273
google-generativeai==0.7.2



numpy<2.0.0
